fix: next round cutoff
