+++ components/user/Sections/Dashboard/Cards/NextRound/NextRound.tsx
@@ -11,24 +11,6 @@ import DashboardCardLoading from '@/components/LoadingModals/DashboardCardLoadin
 
 const fetcher = (url) => fetch(url).then((r) => r.json());
 
-function findCurrentRound(schedules) {
-	const currDate = new Date();
-
-	if (schedules.length) {
-		const nextRound = schedules.filter((round) => {
-			const date = new Date(round.date);
-			if (date < currDate) {
-				return round;
-			}
-		});
-
-		return nextRound[nextRound.length - 1];
-	} else {
-		const nextRound = {};
-		return nextRound;
-	}
-}
-
 export default function NextRound(): JSX.Element {
 	const { data: schedule, error: scheduleError } = useSWR('/api/getSchedule', fetcher);
 
@@ -36,34 +18,9 @@ export default function NextRound(): JSX.Element {
 
 	const [success, setSuccess] = useState(false);
 	const [failure, setFailure] = useState(false);
-	const [nextRound, setNextRound] = useState();
-
-	const [cutOffPast, setCutOffPast] = useState(false);
 
 	if (scheduleError) return <div>Failed to load</div>;
 	if (!schedule) return <DashboardCardLoading />;
-
-	setNextRound(findNextRound(schedule));
-	const currDate = new Date();
-
-	useEffect(() => {
-		const dayOfWeek = currDate.getDay(); //0 is Sunday
-
-		if (dayOfWeek >= 1 && dayOfWeek <= 3) {
-			if (dayOfWeek === 3) {
-				const time = currDate.getHours();
-
-				if (time > 14) {
-					setCutOffPast(false);
-				} else {
-					setNextRound(findCurrentRound(schedule));
-					setCutOffPast(true);
-				}
-			} else {
-				setCutOffPast(true);
-			}
-		}
-	}, []);
 
 	return (
 		<div className='rounded-tl-lg rounded-tr-lg sm:rounded-tr-none relative group bg-white p-6'>
@@ -72,14 +29,8 @@ export default function NextRound(): JSX.Element {
 			<div className='rounded-lg inline-flex p-3 ring-4 ring-white'>
 				<CalendarIcon className='h-6 w-6' aria-hidden='true' />
 			</div>
-			<NextRoundInfo nextRound={nextRound} />
-			<NextRoundForm
-				user={user}
-				setSuccess={setSuccess}
-				setFailure={setFailure}
-				cutOffPast={cutOffPast}
-				nextRound={nextRound}
-			/>
+			<NextRoundInfo schedule={schedule} />
+			<NextRoundForm user={user} setSuccess={setSuccess} setFailure={setFailure} schedule={schedule} />
 		</div>
 	);
 }
